# Speed up dev mode image handling of website