--- a/lib/image-handler.js
+++ b/lib/image-handler.js
@@ -169,7 +169,14 @@ module.exports = function imageHandler(
     // In sync mode, while image is generating in a separate thread, stats are estimated synchronously
     eleventyImage(imgPath, imgOpts)
 
-    const htmlMetadata = eleventyImage.statsSync(imgPath, imgOpts)
+    let htmlMetadata
+    if (isURL(imgSrc)) {
+        throw new Error(
+            `Currently remote images are not supported as they cannot be loaded synchronously, please copy ${imgPath} to the local file system`
+        )
+    } else {
+        htmlMetadata = eleventyImage.statsSync(imgPath, imgOpts)
+    }
 
     return eleventyImage.generateHTML(htmlMetadata, htmlOpts)
 }