--- a/lib/image-handler.js
+++ b/lib/image-handler.js
@@ -78,7 +78,7 @@ module.exports = function imageHandler(
     // Image formats to generate
     // Full list of formats here: https://www.11ty.dev/docs/plugins/image/#output-formats
     // Warning: Avif can be resource-intensive so take care!
-    let formats = ["avif", "webp", "auto"]
+    let formats = ["avif", "webp", "jpeg"]
     let extraOpts = {}
     if (imgSrc.includes(".gif")) {
         formats = ["webp", "gif"]
@@ -88,39 +88,7 @@ module.exports = function imageHandler(
             },
         }
     }
-    const parsedTitle = (imgTitle || '').match(
-        /^(?<skip>@skip ?)?(?<title>.*)/
-    ).groups
-
-    const htmlOpts = {
-        title: parsedTitle.title,
-        alt: imgAlt,
-        loading: 'lazy',
-        decoding: 'async'
-    }
-
-    if (parsedTitle.skip || imgSrc.startsWith('http')) {
-        const options = { ...htmlOpts }
-        const metadata = { img: [{ url: imgSrc }] }
-        return eleventyImage.generateHTML(metadata, options)
-    }
-
-    let formats = ['avif', 'webp', 'jpeg']
-    let extraOpts = {}
-    if (imgSrc.includes('.gif')) {
-        formats = ['webp', 'gif']
-        extraOpts = {
-            sharpOptions: {
-                animated: true
-            },
-        }
-    }
-
-    // Skip slow formats for local development
-    if (DEV_MODE) {
-        formats = formats.filter((format) => !['avif', 'webp'].includes(format))
-    }
-
+    
     // Skip slow formats for local development
     if (DEV_MODE) {
         formats = formats.filter((format) => !['avif', 'webp'].includes(format))
@@ -151,12 +119,23 @@ module.exports = function imageHandler(
         ...extraOpts,
     }
 
+    // Skip image parsing if flag is set (@skip in title)
+    const parsedTitle = (imgTitle || '').match(
+        /^(?<skip>@skip ?)?(?<title>.*)/
+    ).groups
+
     const htmlOpts = {
-        ...(imgTitle && { title: imgTitle }), // skip if null
+        ...(parsedTitle.title && { title: parsedTitle.title }), // skip if null
         alt: imgAlt,
         sizes: htmlSizes.join(", "),
         loading: "lazy",
         decoding: "async",
+    }
+
+    if (parsedTitle.skip || imgSrc.startsWith('http')) {
+        const options = { ...htmlOpts }
+        const metadata = { img: [{ url: imgSrc }] }
+        return eleventyImage.generateHTML(metadata, options)
     }
 
     // In async mode, image is generated, then stats are read and used to generate HTML