## Description

- Reduces cold build time (on my machine) from ~60 seconds to ~50 seconds
- Reduces all follow up builds (on my machine) from ~13 seconds to ~4.5 seconds
- Disables Netlify auto-opening web browser on first build, leading to confusing 404
    -  **Of Note:** This means developers will now need to click on the link (or open the browser) manually
- Disables clearing of caching on rebuild
    -  **Of Note:** This means that on first load, developer might be seeing the old page, manual refresh will be needed

### Possible Follow Ups

- Disable image pipeline entirely locally (speeds up first cold build) - downside possible differences between dev and prod

## Related Issue(s)

https://github.com/flowforge/website/issues/632

## Checklist

<!-- https://flowforge.com/handbook/development/#defining-done -->

 - [x] I have read the [contribution guidelines](https://github.com/flowforge/flowforge/blob/main/CONTRIBUTING.md)
 - [x] I have considered the performance impact of these changes
 - [na] Suitable unit/system level tests have been added and they pass
 - [na] Documentation has been updated