@Pezmc 2023-04-26
> For the practicality of auto-opening, I think we keep this as is. All I'm doing anyway is refreshing the page or clicking the localhost link and seeing the 404 anyway until it's ready, so I think the auto-open is worth keeping.

I disagree here, I think seeing a 404 unexpectedly is confusing for team members who don't know why they're seeing that screen. A user should be waiting for the build to display the finished link, before expecting the page to load.

> I'd like to try this and see what performance improvements we get. Tried to get my head around the pipeline work and workout where to disable, but struggled to find it, so will wait on your input here.

I can do further research here, but let's please move this PR forward in the mean time.