@joepavitt 2023-04-26
> I disagree here, I think seeing a 404 unexpectedly is confusing for team members who don't know why they're seeing that screen. I believe Ian has previously reported it as a bug. Without the auto-open, I'd expect the user to wait for the build to finish, before loading the page.

I believe @iskerrett's concern was more around the amount of time it takes to show the page, rather than the 404 itself, although please correct me if I'm wrong Ian.